+++ src/main.c
@@ -34,6 +34,39 @@
 static const char * TAG = "MAIN";
 
 /***************************
+ * STATIC FUNCTIONS
+****************************/
+
+static void init_wifi_ap(void)
+{
+    char wifi_ap_ssid[32];
+
+#ifdef CONTROLLER_FIRMWARE
+    char *phone = settings_get_phone();
+    if(strcmp(phone, "") == 0)
+    {
+        snprintf(wifi_ap_ssid, 32, "shd_controlador_ñ_configurado");
+    }
+    else
+    {
+        snprintf(wifi_ap_ssid, 32, "shd_controlador");
+    }
+#elif PERIPHERAL_FIRMWARE
+    uint8_t id = settings_get_id();
+    if(id == 0)
+    {
+        snprintf(wifi_ap_ssid, 32, "shd_periférico_ñ_configurado");
+    }
+    else
+    {
+        snprintf(wifi_ap_ssid, 32, "shd_periférico_%d", id);
+    }
+#endif
+    ESP_LOGI(TAG, "Iniciando Access Point %s", wifi_ap_ssid);
+    wifi_ap_init(wifi_ap_ssid, CONFIG_WIFI_AP_PASS);
+}
+
+/***************************
  * MAIN
 ****************************/
 void app_main()
@@ -44,8 +77,6 @@ void app_main()
     ESP_ERROR_CHECK(nvs_flash_init());
     ESP_ERROR_CHECK(esp_netif_init());
     ESP_ERROR_CHECK(esp_event_loop_create_default());
-
-    wifi_ap_init(CONFIG_WIFI_AP_SSID, CONFIG_WIFI_AP_PASS);
 
     #if CONFIG_USE_MDNS
     init_mdns();
@@ -69,6 +100,8 @@ void app_main()
     settings_set_mode(PERIPHERAL);
     settings_update();
 #endif
+wifi_ap_ssid
+    init_wifi_ap();
 
     ESP_LOGI(TAG, "ID: %d - MODE: %s", settings_get_id(), ((uint8_t)settings_get_mode()==1?"CONTROLLER":"PERIPHERAL"));
     protocol_init(((uint8_t)settings_get_mode()==1?CONTROLLER:PERIPHERAL), settings_get_id());