+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+* Updated wifi ssid name depending on the mode and settings;
+
 ### Fixed
 
 ### Removed